Make sure that regenerating the project doesn't invalidate toolchain installs.
Due to the force-download-$tag target, this is correct; if the toolchain
hasn't been downloaded in the first place yet, these stamp files are
irrelevant, but if it has been built already, they will take effect. I
do not know why pre-3.18 CMake invalidates the stamp files when you
regenerate a project, but it does, so here is this stupid commit that
fixes it.