Boot framebuffer console output.
Also various build fixes; fix section expansion for the kernel;
reference libfreestanding.a in a way that makes CMake detect when the .a
file has changed (outside of the library's own build system). That last
one will probably require some helper functions to future-proof before
the use of the current pattern gets out of hand.