Refactor the PMM, make UEFI loader mark BSP stack as not free.
The refactored PMM is now fully parametrized on the page sizes supported
by the arch the kernel is being compiled for, as opposed to hardcoding
the x86 page sizes. This may or may not be useful in the future, but is
definitely cleaner now, also in the PMM interface.

Trying to allocate more than a page of stack space for APs has revealed
a bug, where the active stack (inherited from the UEFI loader) was
treated as free frames, and pushed onto the frame stack. I'm rather glad
that this has revealed itself now, and not in some more arcane manner
much, much later in this project. This has now been fixed; the UEFI
loader now detects this and marks the stack frames as a special memory
region. It now also does a canonicalization pass over the memory map
before passing it on, because it was necessary for the sanity of the
code that marks the stack region. The kernel PMM will continue
containing a canonicalization pass, in case doing it in the loader is
more difficult for loaders other than UEFI.