Reach userspace.
The current scheduler is still very much a heavy stub; it only has a
single queue of threads to run; it always selects the one that had CPU
time the longest time ago; and finally, it doesn't actually do any
preemption yet. But it's enough to reach userspace the way I want the
kernel to reach userspace initially, so it's deemed Good Enough for a
commit.

Next steps involve:

1) Creating the nested initrd. (tar? unsure yet)
2) Writing code to "parse" the nested initrd inside the bootinit
process.
3) Build a stub service, but as an ELF file this time, not a flat
binary.
4) Write an ELF parsing library.
5) Make bootinit create a process which will find the stub service in
the initrd, parse its executable file, create a process out of it, and
start it (and verify communication over IPC).

There's a *lot* of intermediate things that need to happen for the above
to become reality, but that's the overall plan for the next steps.