Kernel-private VDSO, bootinit, VMOs, beginnings of the object/handle/token model.
I believe that this is *roughly* the final general shape of kernel_main.
There's probably more things that need to be sent to the bootinit
process, but that's fine.

Also introduce an AVL tree datastructure that is used in a couple of
places, std::optional, more code for iterator concepts (that ends up not
being actually needed here), and a fix to chained_allocator (...the
previous code stupidly only used a single slot per allocated page,
uhhh.)