Initial implementation of ELF loading and linking.
This commit contains a toolchain crime - it makes the toolchain files
pretend that the target system is a Linux. This is because Clang's
driver is doing something extremely stupid and invoking the linking step
*through GCC* when it doesn't recognize the target, and it also does it
*incorrectly*, because --sysroot gets lost somewhere. Because we now
need to *correctly* link some shared objects, and because doing it
through ld.lld directly is much more of a pain than this hack, we'll do
the hack for this. *Before this is merged*, however, the branch should
also receive a commit that patches LLVM to understand
*-reaveros-{elf,kernel} target triples, but that is going to be way more
involved than I have time, energy, and space in this commit for.

This commit also modifies how blocking syscalls lock memory ranges; the
locks no longer persist across suspensions. This is because I'm also
adding a mapping unmap syscall which needs a unique ownership of a
mapping, and it was either this or making unmap itself blocking, and
making mappings have thread queues on their locks. I chose the simpler
approach. Also also, the mapping lock type is now a shared mutex, as it
should've been from the start.

Finally - the first actual userspace process will be a logger, not a
vasmgr. This is to bring the number of binaries that need to contain the
"send some pointers over a mailbox to a kernel thread" logging mechanism
to its bare minimum. This way, all the other processes will start by
immediately using the logger IPC protocol, which will make them simpler
in this regard.