Coerce llvm's build system to do the thing I'm asking it to do.
So... there must be *some* runtimes included on the main runtimes list
to trigger the multiconfig thing for runtimes, but it doesn't seem to
really matter what's on that list, just... anything. BUT if compiler-rt
is included, some additional magical targets for builtins are *also*
included, and there does not exist a way to turn them off - they get
included as long as LLVM_BUILTIN_TARGETS is empty, but I can't set it to
a bogus list or it goes haywire. So I just lie to the build system
saying I won't be building compiler-rt, so that whole section of it is
skipped, but then enable compiler-rt on the per-target basis anyway.

Very deep sigh.