Fix the behavior of ExternalProject in newer versions of cmake.
Something after 3.18 has broken the way I was doing ExternalProject in a
parallel build; if both foo and foo-install were dependent on, cmake
would run all the steps twice in parallel, breaking everything.

I originally "fixed" it by switching policy CMP0114, but that was wrong,
because it made targets be ran much more eagerly, and made commands
attached to an install step target (I think) always run, which is
incorrect and touched too many files (and I noticed it only because I
was reinstalling libclang_rt.builtin.a from the wrong location for a
time).

*Now* I fixed it by being consistent in which targets I depend on: for
internal components of the OS, I now always depend on the full target
(so `foo` instead of `foo-install`), while for ExternalProjects that are
not meant to be rebuilt every time, i.e. the toolchain pieces, I now
always depend on the `foo-install` targets (similarly to how it was
earlier), and this coerces cmake to not run the same step twice in
parallel, so parallel builds work again.

I hope this is the last time I need to touch this magic.