Correctly handle elf dependency and init orders, print from logger.
This does a bunch of refactoring of how the code is structured. The
bootinit printing class is now reused by the logger (and logger only)
for the purpose of having a bootstrap print target for the steps the
logger takes. I considered making it its own library, but it depends on
the kernel generating the syscall headers, and is used by what is
currently a part of the kernel build, so I just hacked together an out
of tree reference to the files from the logger, and this works fine.

Additionally, all the bootinit specific ELF handling is now in its own
file; it correctly processes symbols definition in the right dependency
order; and it now also computes and provides init_array data (with most
of the code for preinit and fini also in place), and a startup function
in rosert can read that and execute it in the correct order when the
information is present.

Finally, the logger now actually prints text.

There's a number of bug fixes also included in this commit, too numerous
to actually list.

This also includes an implementation of std::rotate, which currently is
not actually tested (and just used by the elf handling machinery). I'll
leave it as a future TODO to actually add tests for that, when I'll go
and start adding tests for more of the stdlib features. The machinery is
all there, but I'll write the tests when I run out of easy things to do
for a while.