Resolve a prior circular dependency between component.
Because REAVEROS_COMPONENT_DEPENDS required just a list of targets to
always depend on, rosestd needed to always depend on rosert, which in
turn needs to now depend on the kernel (in form of all-kernels, because
the kernel target itself does not yet exist when libraries are processed
- this is slightly inconvenient, but should be fine). This meant that if
rosert was made to *actually* depend on all-kernels, there would be a
circular dependency, because the kernel depends on rosestd:

rosert -> all-kernels -> kernel-$arch -> rosestd -> rosert

Now, there is a special option for libraries to only depend on another
component in hosted mode (this will possibly need to be extended to
testing mode, but I will only do that when necessary); the freestanding
rosestd doesn't depend on rosert (logically), so this new feature allows
it to not depend on rosert at all. So the dependency chain is now as
follows:

rosestd-hosted -> rosert-hosted -> all-kernels -> rosestd-freestanding

This also allows removing the noop hackery from rosert's build system
for the freestanding and testing modes.