Actually schedule work on more than one core.
This is a really dumb algorithm as it is, but the skeleton for how I
want hyperthreading / NUMA support code to be structured is now in
place. I'll start *actually* handling them at a later date.

I am still not quite sure how to deal with big.LITTLE, but we'll get
there.